Refactoring: Loading Product Data from a JSON File
I changed the way product data is managed in JavaScript. Instead of using an array in the code, I'm now loading product data from an external JSON file.

I created a JSON file that contains the product data. This makes the code more organized and allows product data to be easily updated or modified separately.

I refactored the exibirmaisprocurados function to load product data from the JSON file. Now, the function reads data from the JSON file and dynamically creates product elements.

Furthermore, I ensured that the exibirmaisprocurados function is imported and reloaded every time the section-switch function reloads the page. This ensures that the products are loaded correctly, even when the user returns to the home page.