The 8 Most Clicked Products from JSON Data
In this commit, I implemented a feature that loads the 8 most clicked products from the JSON data. The products are now sorted in descending order based on the number of clicks, and the top 8 products are selected for display. This change enhances the user experience by highlighting the most popular products on the page.

Note: Please note that users are currently unable to add clicks to the JSON data in this version.